Move musicvideos and series
