+++ resources/lib/itemtypes.py
@@ -18,7 +18,7 @@ import embydb_functions as embydb
 import kodidb_functions as kodidb
 import read_embyserver as embyserver
 import musicutils
-from objects import Movies
+from objects import Movies, MusicVideos, TVShows
 from utils import window, settings, language as lang, kodiSQL
 
 #################################################################################################
@@ -62,6 +62,7 @@ class Items(object):
 
             'Movie': Movies,
             'BoxSet': Movies,
+            'MusicVideo': MusicVideos,
             'Series': TVShows,
             'Season': TVShows,
             'Episode': TVShows,
@@ -227,1050 +228,6 @@ class Items(object):
                     time=time,
                     sound=False)
 
-
-class MusicVideos(Items):
-
-    
-    def __init__(self, embycursor, kodicursor):
-        Items.__init__(self, embycursor, kodicursor)
-
-    def added(self, items, pdialog):
-
-        total = len(items)
-        count = 0
-        for mvideo in items:
-
-            title = mvideo['Name']
-            if pdialog:
-                percentage = int((float(count) / float(total))*100)
-                pdialog.update(percentage, message=title)
-                count += 1
-            self.add_update(mvideo)
-            if not pdialog and self.contentmsg:
-                self.contentPop(title, self.newvideo_time)
-
-
-    def add_update(self, item, viewtag=None, viewid=None):
-        # Process single music video
-        kodicursor = self.kodicursor
-        emby_db = self.emby_db
-        artwork = self.artwork
-        API = api.API(item)
-
-        if item.get('LocationType') == "Virtual": # TODO: Filter via api instead
-            log.info("Skipping virtual episode: %s", item['Name'])
-            return
-
-        # If the item already exist in the local Kodi DB we'll perform a full item update
-        # If the item doesn't exist, we'll add it to the database
-        update_item = True
-        itemid = item['Id']
-        emby_dbitem = emby_db.getItem_byId(itemid)
-        try:
-            mvideoid = emby_dbitem[0]
-            fileid = emby_dbitem[1]
-            pathid = emby_dbitem[2]
-            log.info("mvideoid: %s fileid: %s pathid: %s" % (mvideoid, fileid, pathid))
-        
-        except TypeError:
-            update_item = False
-            log.debug("mvideoid: %s not found." % itemid)
-            # mvideoid
-            kodicursor.execute("select coalesce(max(idMVideo),0) from musicvideo")
-            mvideoid = kodicursor.fetchone()[0] + 1
-
-        else:
-            # Verification the item is still in Kodi
-            query = "SELECT * FROM musicvideo WHERE idMVideo = ?"
-            kodicursor.execute(query, (mvideoid,))
-            try:
-                kodicursor.fetchone()[0]
-            except TypeError:
-                # item is not found, let's recreate it.
-                update_item = False
-                log.info("mvideoid: %s missing from Kodi, repairing the entry." % mvideoid)
-
-        if not viewtag or not viewid:
-            # Get view tag from emby
-            viewtag, viewid, mediatype = self.emby.getView_embyId(itemid)
-            log.debug("View tag found: %s" % viewtag)
-
-        # fileId information
-        checksum = API.get_checksum()
-        dateadded = API.get_date_created()
-        userdata = API.get_userdata()
-        playcount = userdata['PlayCount']
-        dateplayed = userdata['LastPlayedDate']
-
-        # item details
-        runtime = API.get_runtime()
-        plot = API.get_overview()
-        title = item['Name']
-        year = item.get('ProductionYear')
-        genres = item['Genres']
-        genre = " / ".join(genres)
-        studios = API.get_studios()
-        studio = " / ".join(studios)
-        artist = " / ".join(item.get('Artists'))
-        album = item.get('Album')
-        track = item.get('Track')
-        people = API.get_people()
-        director = " / ".join(people['Director'])
-
-        
-        ##### GET THE FILE AND PATH #####
-        playurl = API.get_file_path()
-
-        if "\\" in playurl:
-            # Local path
-            filename = playurl.rsplit("\\", 1)[1]
-        else: # Network share
-            filename = playurl.rsplit("/", 1)[1]
-
-        if self.directpath:
-            # Direct paths is set the Kodi way
-            if not self.pathValidation(playurl):
-                return False
-            
-            path = playurl.replace(filename, "")
-            window('emby_pathverified', value="true")
-        else:
-            # Set plugin path and media flags using real filename
-            path = "plugin://plugin.video.emby.musicvideos/"
-            params = {
-
-                'filename': filename.encode('utf-8'),
-                'id': itemid,
-                'dbid': mvideoid,
-                'mode': "play"
-            }
-            filename = "%s?%s" % (path, urllib.urlencode(params))
-
-
-        ##### UPDATE THE MUSIC VIDEO #####
-        if update_item:
-            log.info("UPDATE mvideo itemid: %s - Title: %s" % (itemid, title))
-            
-            # Update path
-            query = "UPDATE path SET strPath = ? WHERE idPath = ?"
-            kodicursor.execute(query, (path, pathid))
-
-            # Update the filename
-            query = "UPDATE files SET strFilename = ?, dateAdded = ? WHERE idFile = ?"
-            kodicursor.execute(query, (filename, dateadded, fileid))
-
-            # Update the music video entry
-            query = ' '.join((
-                
-                "UPDATE musicvideo",
-                "SET c00 = ?, c04 = ?, c05 = ?, c06 = ?, c07 = ?, c08 = ?, c09 = ?, c10 = ?,",
-                    "c11 = ?, c12 = ?"
-                "WHERE idMVideo = ?"
-            ))
-            kodicursor.execute(query, (title, runtime, director, studio, year, plot, album,
-                artist, genre, track, mvideoid))
-
-            # Update the checksum in emby table
-            emby_db.updateReference(itemid, checksum)
-        
-        ##### OR ADD THE MUSIC VIDEO #####
-        else:
-            log.info("ADD mvideo itemid: %s - Title: %s" % (itemid, title))
-            
-            # Add path
-            query = ' '.join((
-
-                "SELECT idPath",
-                "FROM path",
-                "WHERE strPath = ?"
-            ))
-            kodicursor.execute(query, (path,))
-            try:
-                pathid = kodicursor.fetchone()[0]
-            except TypeError:
-                kodicursor.execute("select coalesce(max(idPath),0) from path")
-                pathid = kodicursor.fetchone()[0] + 1
-                query = (
-                    '''
-                    INSERT OR REPLACE INTO path(
-                        idPath, strPath, strContent, strScraper, noUpdate)
-
-                    VALUES (?, ?, ?, ?, ?)
-                    '''
-                )
-                kodicursor.execute(query, (pathid, path, "musicvideos", "metadata.local", 1))
-
-            # Add the file
-            kodicursor.execute("select coalesce(max(idFile),0) from files")
-            fileid = kodicursor.fetchone()[0] + 1
-            query = (
-                '''
-                INSERT INTO files(
-                    idFile, idPath, strFilename, dateAdded)
-
-                VALUES (?, ?, ?, ?)
-                '''
-            )
-            kodicursor.execute(query, (fileid, pathid, filename, dateadded))
-            
-            # Create the musicvideo entry
-            query = (
-                '''
-                INSERT INTO musicvideo(
-                    idMVideo, idFile, c00, c04, c05, c06, c07, c08, c09, c10, c11, c12)
-
-                VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                '''
-            )
-            kodicursor.execute(query, (mvideoid, fileid, title, runtime, director, studio,
-                year, plot, album, artist, genre, track))
-
-            # Create the reference in emby table
-            emby_db.addReference(itemid, mvideoid, "MusicVideo", "musicvideo", fileid, pathid,
-                checksum=checksum, mediafolderid=viewid)
-
-        
-        # Process cast
-        people = item['People']
-        artists = item['ArtistItems']
-        for artist in artists:
-            artist['Type'] = "Artist"
-        people.extend(artists)
-        people = artwork.get_people_artwork(people)
-        self.kodi_db.addPeople(mvideoid, people, "musicvideo")
-        # Process genres
-        self.kodi_db.addGenres(mvideoid, genres, "musicvideo")
-        # Process artwork
-        artwork.add_artwork(artwork.get_all_artwork(item), mvideoid, "musicvideo", kodicursor)
-        # Process stream details
-        streams = API.get_media_streams()
-        self.kodi_db.addStreams(fileid, streams, runtime)
-        # Process studios
-        self.kodi_db.addStudios(mvideoid, studios, "musicvideo")
-        # Process tags: view, emby tags
-        tags = [viewtag]
-        tags.extend(item['Tags'])
-        if userdata['Favorite']:
-            tags.append("Favorite musicvideos")
-        self.kodi_db.addTags(mvideoid, tags, "musicvideo")
-        # Process playstates
-        resume = API.adjust_resume(userdata['Resume'])
-        total = round(float(runtime), 6)
-        self.kodi_db.addPlaystate(fileid, resume, total, playcount, dateplayed)
-
-    def updateUserdata(self, item):
-        # This updates: Favorite, LastPlayedDate, Playcount, PlaybackPositionTicks
-        # Poster with progress bar
-        emby_db = self.emby_db
-        API = api.API(item)
-        
-        # Get emby information
-        itemid = item['Id']
-        checksum = API.get_checksum()
-        userdata = API.get_userdata()
-        runtime = API.get_runtime()
-
-        # Get Kodi information
-        emby_dbitem = emby_db.getItem_byId(itemid)
-        try:
-            mvideoid = emby_dbitem[0]
-            fileid = emby_dbitem[1]
-            log.info(
-                "Update playstate for musicvideo: %s fileid: %s"
-                % (item['Name'], fileid))
-        except TypeError:
-            return
-
-        # Process favorite tags
-        if userdata['Favorite']:
-            self.kodi_db.addTag(mvideoid, "Favorite musicvideos", "musicvideo")
-        else:
-            self.kodi_db.removeTag(mvideoid, "Favorite musicvideos", "musicvideo")
-
-        # Process playstates
-        playcount = userdata['PlayCount']
-        dateplayed = userdata['LastPlayedDate']
-        resume = API.adjust_resume(userdata['Resume'])
-        total = round(float(runtime), 6)
-
-        self.kodi_db.addPlaystate(fileid, resume, total, playcount, dateplayed)
-        emby_db.updateReference(itemid, checksum)
-
-    def remove(self, itemid):
-        # Remove mvideoid, fileid, pathid, emby reference
-        emby_db = self.emby_db
-        kodicursor = self.kodicursor
-        artwork = self.artwork
-
-        emby_dbitem = emby_db.getItem_byId(itemid)
-        try:
-            mvideoid = emby_dbitem[0]
-            fileid = emby_dbitem[1]
-            pathid = emby_dbitem[2]
-            log.info("Removing mvideoid: %s fileid: %s" % (mvideoid, fileid, pathid))
-        except TypeError:
-            return
-
-        # Remove artwork
-        query = ' '.join((
-
-            "SELECT url, type",
-            "FROM art",
-            "WHERE media_id = ?",
-            "AND media_type = 'musicvideo'"
-        ))
-        kodicursor.execute(query, (mvideoid,))
-        for row in kodicursor.fetchall():
-            
-            url = row[0]
-            imagetype = row[1]
-            if imagetype in ("poster", "fanart"):
-                artwork.delete_cached_artwork(url)
-
-        kodicursor.execute("DELETE FROM musicvideo WHERE idMVideo = ?", (mvideoid,))
-        kodicursor.execute("DELETE FROM files WHERE idFile = ?", (fileid,))
-        if self.directpath:
-            kodicursor.execute("DELETE FROM path WHERE idPath = ?", (pathid,))
-        self.embycursor.execute("DELETE FROM emby WHERE emby_id = ?", (itemid,))
-
-        log.info("Deleted musicvideo %s from kodi database" % itemid)
-
-class TVShows(Items):
-
-
-    def __init__(self, embycursor, kodicursor):
-        Items.__init__(self, embycursor, kodicursor)
-
-    def added(self, items, pdialog):
-        
-        total = len(items)
-        count = 0
-        for tvshow in items:
-
-            title = tvshow['Name']
-            if pdialog:
-                percentage = int((float(count) / float(total))*100)
-                pdialog.update(percentage, message=title)
-                count += 1
-            self.add_update(tvshow)
-            # Add episodes
-            all_episodes = self.emby.getEpisodesbyShow(tvshow['Id'])
-            self.added_episode(all_episodes['Items'], pdialog)
-
-    def added_season(self, items, pdialog):
-        
-        total = len(items)
-        count = 0
-        for season in items:
-
-            title = "%s - %s" % (season.get('SeriesName', "Unknown"), season['Name'])
-            if pdialog:
-                percentage = int((float(count) / float(total))*100)
-                pdialog.update(percentage, message=title)
-                count += 1
-            self.add_updateSeason(season)
-            # Add episodes
-            all_episodes = self.emby.getEpisodesbySeason(season['Id'])
-            self.added_episode(all_episodes['Items'], pdialog)
-
-    def added_episode(self, items, pdialog):
-        
-        total = len(items)
-        count = 0
-        for episode in items:
-            title = "%s - %s" % (episode.get('SeriesName', "Unknown"), episode['Name'])
-            if pdialog:
-                percentage = int((float(count) / float(total))*100)
-                pdialog.update(percentage, message=title)
-                count += 1
-            self.add_updateEpisode(episode)
-            if not pdialog and self.contentmsg:
-                self.contentPop(title, self.newvideo_time)
-
-
-    def add_update(self, item, viewtag=None, viewid=None):
-        # Process single tvshow
-        kodicursor = self.kodicursor
-        emby = self.emby
-        emby_db = self.emby_db
-        artwork = self.artwork
-        API = api.API(item)
-
-        if settings('syncEmptyShows') == "false" and not item.get('RecursiveItemCount'):
-            log.info("Skipping empty show: %s" % item['Name'])
-            return
-        # If the item already exist in the local Kodi DB we'll perform a full item update
-        # If the item doesn't exist, we'll add it to the database
-        update_item = True
-        force_episodes = False
-        itemid = item['Id']
-        emby_dbitem = emby_db.getItem_byId(itemid)
-        try:
-            showid = emby_dbitem[0]
-            pathid = emby_dbitem[2]
-            log.info("showid: %s pathid: %s" % (showid, pathid))
-        
-        except TypeError:
-            update_item = False
-            log.debug("showid: %s not found." % itemid)
-            kodicursor.execute("select coalesce(max(idShow),0) from tvshow")
-            showid = kodicursor.fetchone()[0] + 1
-
-        else:
-            # Verification the item is still in Kodi
-            query = "SELECT * FROM tvshow WHERE idShow = ?"
-            kodicursor.execute(query, (showid,))
-            try:
-                kodicursor.fetchone()[0]
-            except TypeError:
-                # item is not found, let's recreate it.
-                update_item = False
-                log.info("showid: %s missing from Kodi, repairing the entry." % showid)
-                # Force re-add episodes after the show is re-created.
-                force_episodes = True
-
-
-        if viewtag is None or viewid is None:
-            # Get view tag from emby
-            viewtag, viewid, mediatype = emby.getView_embyId(itemid)
-            log.debug("View tag found: %s" % viewtag)
-
-        # fileId information
-        checksum = API.get_checksum()
-        dateadded = API.get_date_created()
-        userdata = API.get_userdata()
-        playcount = userdata['PlayCount']
-        dateplayed = userdata['LastPlayedDate']
-
-        # item details
-        genres = item['Genres']
-        title = item['Name']
-        plot = API.get_overview()
-        rating = item.get('CommunityRating')
-        premieredate = API.get_premiere_date()
-        tvdb = API.get_provider('Tvdb')
-        sorttitle = item['SortName']
-        mpaa = API.get_mpaa()
-        genre = " / ".join(genres)
-        studios = API.get_studios()
-        studio = " / ".join(studios)
-
-        # Verify series pooling
-        if not update_item and tvdb:
-            query = "SELECT idShow FROM tvshow WHERE C12 = ?"
-            kodicursor.execute(query, (tvdb,))
-            try:
-                temp_showid = kodicursor.fetchone()[0]
-            except TypeError:
-                pass
-            else:
-                emby_other = emby_db.getItem_byKodiId(temp_showid, "tvshow")
-                if emby_other and viewid == emby_other[2]:
-                    log.info("Applying series pooling for %s", title)
-                    emby_other_item = emby_db.getItem_byId(emby_other[0])
-                    showid = emby_other_item[0]
-                    pathid = emby_other_item[2]
-                    log.info("showid: %s pathid: %s" % (showid, pathid))
-                    # Create the reference in emby table
-                    emby_db.addReference(itemid, showid, "Series", "tvshow", pathid=pathid,
-                                        checksum=checksum, mediafolderid=viewid)
-                    update_item = True
-
-        
-        ##### GET THE FILE AND PATH #####
-        playurl = API.get_file_path()
-
-        if self.directpath:
-            # Direct paths is set the Kodi way
-            if "\\" in playurl:
-                # Local path
-                path = "%s\\" % playurl
-                toplevelpath = "%s\\" % dirname(dirname(path))
-            else:
-                # Network path
-                path = "%s/" % playurl
-                toplevelpath = "%s/" % dirname(dirname(path))
-
-            if not self.pathValidation(path):
-                return False
-
-            window('emby_pathverified', value="true")
-        else:
-            # Set plugin path
-            toplevelpath = "plugin://plugin.video.emby.tvshows/"
-            path = "%s%s/" % (toplevelpath, itemid)
-
-
-        ##### UPDATE THE TVSHOW #####
-        if update_item:
-            log.info("UPDATE tvshow itemid: %s - Title: %s" % (itemid, title))
-
-            # Update the tvshow entry
-            query = ' '.join((
-                
-                "UPDATE tvshow",
-                "SET c00 = ?, c01 = ?, c04 = ?, c05 = ?, c08 = ?, c09 = ?,",
-                    "c12 = ?, c13 = ?, c14 = ?, c15 = ?",
-                "WHERE idShow = ?"
-            ))
-            kodicursor.execute(query, (title, plot, rating, premieredate, genre, title,
-                tvdb, mpaa, studio, sorttitle, showid))
-
-            # Update the checksum in emby table
-            emby_db.updateReference(itemid, checksum)
-        
-        ##### OR ADD THE TVSHOW #####
-        else:
-            log.info("ADD tvshow itemid: %s - Title: %s" % (itemid, title))
-            
-            # Add top path
-            toppathid = self.kodi_db.addPath(toplevelpath)
-            query = ' '.join((
-
-                "UPDATE path",
-                "SET strPath = ?, strContent = ?, strScraper = ?, noUpdate = ?",
-                "WHERE idPath = ?"
-            ))
-            kodicursor.execute(query, (toplevelpath, "tvshows", "metadata.local", 1, toppathid))
-            
-            # Add path
-            pathid = self.kodi_db.addPath(path)
-            
-            # Create the tvshow entry
-            query = (
-                '''
-                INSERT INTO tvshow(
-                    idShow, c00, c01, c04, c05, c08, c09, c12, c13, c14, c15) 
-
-                VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                '''
-            )
-            kodicursor.execute(query, (showid, title, plot, rating, premieredate, genre,
-                title, tvdb, mpaa, studio, sorttitle))
-
-            # Link the path
-            query = "INSERT INTO tvshowlinkpath(idShow, idPath) values(?, ?)"
-            kodicursor.execute(query, (showid, pathid))
-
-            # Create the reference in emby table
-            emby_db.addReference(itemid, showid, "Series", "tvshow", pathid=pathid,
-                                checksum=checksum, mediafolderid=viewid)
-
-        # Update the path
-        query = ' '.join((
-
-            "UPDATE path",
-            "SET strPath = ?, strContent = ?, strScraper = ?, noUpdate = ?",
-            "WHERE idPath = ?"
-        ))
-        kodicursor.execute(query, (path, None, None, 1, pathid))
-        
-        # Process cast
-        people = artwork.get_people_artwork(item['People'])
-        self.kodi_db.addPeople(showid, people, "tvshow")
-        # Process genres
-        self.kodi_db.addGenres(showid, genres, "tvshow")
-        # Process artwork
-        artwork.add_artwork(artwork.get_all_artwork(item), showid, "tvshow", kodicursor)
-        # Process studios
-        self.kodi_db.addStudios(showid, studios, "tvshow")
-        # Process tags: view, emby tags
-        tags = [viewtag]
-        tags.extend(item['Tags'])
-        if userdata['Favorite']:
-            tags.append("Favorite tvshows")
-        self.kodi_db.addTags(showid, tags, "tvshow")
-        # Process seasons
-        all_seasons = emby.getSeasons(itemid)
-        for season in all_seasons['Items']:
-            self.add_updateSeason(season, showid=showid)
-        else:
-            # Finally, refresh the all season entry
-            seasonid = self.kodi_db.addSeason(showid, -1)
-            # Process artwork
-            artwork.add_artwork(artwork.get_all_artwork(item), seasonid, "season", kodicursor)
-
-        if force_episodes:
-            # We needed to recreate the show entry. Re-add episodes now.
-            log.info("Repairing episodes for showid: %s %s" % (showid, title))
-            all_episodes = emby.getEpisodesbyShow(itemid)
-            self.added_episode(all_episodes['Items'], None)
-
-    def add_updateSeason(self, item, showid=None):
-
-        kodicursor = self.kodicursor
-        emby_db = self.emby_db
-        artwork = self.artwork
-
-        seasonnum = item.get('IndexNumber', 1)
-
-        if showid is None:
-            try:
-                seriesId = item['SeriesId']
-                showid = emby_db.getItem_byId(seriesId)[0]
-            except KeyError:
-                return
-            except TypeError:
-                # Show is missing, update show instead.
-                show = self.emby.getItem(seriesId)
-                self.add_update(show)
-                return
-        
-        seasonid = self.kodi_db.addSeason(showid, seasonnum, item['Name'])
-        
-        if item['LocationType'] != "Virtual":
-            # Create the reference in emby table
-            emby_db.addReference(item['Id'], seasonid, "Season", "season", parentid=showid)
-
-        # Process artwork
-        artwork.add_artwork(artwork.get_all_artwork(item), seasonid, "season", kodicursor)
-
-    def add_updateEpisode(self, item):
-        # Process single episode
-        kodicursor = self.kodicursor
-        emby_db = self.emby_db
-        artwork = self.artwork
-        API = api.API(item)
-
-        if item.get('LocationType') == "Virtual": # TODO: Filter via api instead
-            log.info("Skipping virtual episode: %s", item['Name'])
-            return
-
-        # If the item already exist in the local Kodi DB we'll perform a full item update
-        # If the item doesn't exist, we'll add it to the database
-        update_item = True
-        itemid = item['Id']
-        emby_dbitem = emby_db.getItem_byId(itemid)
-        try:
-            episodeid = emby_dbitem[0]
-            fileid = emby_dbitem[1]
-            pathid = emby_dbitem[2]
-            log.info("episodeid: %s fileid: %s pathid: %s" % (episodeid, fileid, pathid))
-        
-        except TypeError:
-            update_item = False
-            log.debug("episodeid: %s not found." % itemid)
-            # episodeid
-            kodicursor.execute("select coalesce(max(idEpisode),0) from episode")
-            episodeid = kodicursor.fetchone()[0] + 1
-
-        else:
-            # Verification the item is still in Kodi
-            query = "SELECT * FROM episode WHERE idEpisode = ?"
-            kodicursor.execute(query, (episodeid,))
-            try:
-                kodicursor.fetchone()[0]
-            except TypeError:
-                # item is not found, let's recreate it.
-                update_item = False
-                log.info("episodeid: %s missing from Kodi, repairing the entry." % episodeid)
-
-        # fileId information
-        checksum = API.get_checksum()
-        dateadded = API.get_date_created()
-        userdata = API.get_userdata()
-        playcount = userdata['PlayCount']
-        dateplayed = userdata['LastPlayedDate']
-
-        # item details
-        people = API.get_people()
-        writer = " / ".join(people['Writer'])
-        director = " / ".join(people['Director'])
-        title = item['Name']
-        plot = API.get_overview()
-        rating = item.get('CommunityRating')
-        runtime = API.get_runtime()
-        premieredate = API.get_premiere_date()
-
-        # episode details
-        try:
-            seriesId = item['SeriesId']
-        except KeyError:
-            # Missing seriesId, skip
-            log.error("Skipping: %s. SeriesId is missing." % itemid)
-            return False
-            
-        season = item.get('ParentIndexNumber')
-        episode = item.get('IndexNumber', -1)
-       
-        if season is None:
-            if item.get('AbsoluteEpisodeNumber'):
-                # Anime scenario
-                season = 1
-                episode = item['AbsoluteEpisodeNumber']
-            else:
-                season = -1 if "Specials" not in item['Path'] else 0
-
-        # Specials ordering within season
-        if item.get('AirsAfterSeasonNumber'):
-            airsBeforeSeason = item['AirsAfterSeasonNumber']
-            airsBeforeEpisode = 4096 # Kodi default number for afterseason ordering
-        else:
-            airsBeforeSeason = item.get('AirsBeforeSeasonNumber')
-            airsBeforeEpisode = item.get('AirsBeforeEpisodeNumber')
-
-        # Append multi episodes to title
-        if item.get('IndexNumberEnd'):              
-            title = "| %02d | %s" % (item['IndexNumberEnd'], title)
-
-        # Get season id
-        show = emby_db.getItem_byId(seriesId)
-        try:
-            showid = show[0]
-        except TypeError:
-            # Show is missing from database
-            show = self.emby.getItem(seriesId)
-            self.add_update(show)
-            show = emby_db.getItem_byId(seriesId)
-            try:
-                showid = show[0]
-            except TypeError:
-                log.error("Skipping: %s. Unable to add series: %s." % (itemid, seriesId))
-                return False
-
-        seasonid = self.kodi_db.addSeason(showid, season)
-
-        
-        ##### GET THE FILE AND PATH #####
-        playurl = API.get_file_path()
-
-        if "\\" in playurl:
-            # Local path
-            filename = playurl.rsplit("\\", 1)[1]
-        else: # Network share
-            filename = playurl.rsplit("/", 1)[1]
-
-        if self.directpath:
-            # Direct paths is set the Kodi way
-            if not self.pathValidation(playurl):
-                return False
-            
-            path = playurl.replace(filename, "")
-            window('emby_pathverified', value="true")
-        else:
-            # Set plugin path and media flags using real filename
-            path = "plugin://plugin.video.emby.tvshows/%s/" % seriesId
-            params = {
-
-                'filename': filename.encode('utf-8'),
-                'id': itemid,
-                'dbid': episodeid,
-                'mode': "play"
-            }
-            filename = "%s?%s" % (path, urllib.urlencode(params))
-
-
-        ##### UPDATE THE EPISODE #####
-        if update_item:
-            log.info("UPDATE episode itemid: %s - Title: %s" % (itemid, title))
-
-            # Update the movie entry
-            if self.kodiversion in (16, 17):
-                # Kodi Jarvis, Krypton
-                query = ' '.join((
-                
-                    "UPDATE episode",
-                    "SET c00 = ?, c01 = ?, c03 = ?, c04 = ?, c05 = ?, c09 = ?, c10 = ?,",
-                        "c12 = ?, c13 = ?, c14 = ?, c15 = ?, c16 = ?, idSeason = ?, idShow = ?",
-                    "WHERE idEpisode = ?"
-                ))
-                kodicursor.execute(query, (title, plot, rating, writer, premieredate,
-                    runtime, director, season, episode, title, airsBeforeSeason,
-                    airsBeforeEpisode, seasonid, showid, episodeid))
-            else:
-                query = ' '.join((
-                    
-                    "UPDATE episode",
-                    "SET c00 = ?, c01 = ?, c03 = ?, c04 = ?, c05 = ?, c09 = ?, c10 = ?,",
-                        "c12 = ?, c13 = ?, c14 = ?, c15 = ?, c16 = ?, idShow = ?",
-                    "WHERE idEpisode = ?"
-                ))
-                kodicursor.execute(query, (title, plot, rating, writer, premieredate,
-                    runtime, director, season, episode, title, airsBeforeSeason,
-                    airsBeforeEpisode, showid, episodeid))
-
-            # Update the checksum in emby table
-            emby_db.updateReference(itemid, checksum)
-            # Update parentid reference
-            emby_db.updateParentId(itemid, seasonid)
-        
-        ##### OR ADD THE EPISODE #####
-        else:
-            log.info("ADD episode itemid: %s - Title: %s" % (itemid, title))
-            
-            # Add path
-            pathid = self.kodi_db.addPath(path)
-            # Add the file
-            fileid = self.kodi_db.addFile(filename, pathid)
-            
-            # Create the episode entry
-            if self.kodiversion in (16, 17):
-                # Kodi Jarvis, Krypton
-                query = (
-                    '''
-                    INSERT INTO episode(
-                        idEpisode, idFile, c00, c01, c03, c04, c05, c09, c10, c12, c13, c14,
-                        idShow, c15, c16, idSeason)
-
-                    VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                    '''
-                )
-                kodicursor.execute(query, (episodeid, fileid, title, plot, rating, writer,
-                    premieredate, runtime, director, season, episode, title, showid,
-                    airsBeforeSeason, airsBeforeEpisode, seasonid))
-            else:
-                query = (
-                    '''
-                    INSERT INTO episode(
-                        idEpisode, idFile, c00, c01, c03, c04, c05, c09, c10, c12, c13, c14,
-                        idShow, c15, c16)
-
-                    VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                    '''
-                )
-                kodicursor.execute(query, (episodeid, fileid, title, plot, rating, writer,
-                    premieredate, runtime, director, season, episode, title, showid,
-                    airsBeforeSeason, airsBeforeEpisode))
-
-            # Create the reference in emby table
-            emby_db.addReference(itemid, episodeid, "Episode", "episode", fileid, pathid,
-                seasonid, checksum)
-
-        # Update the path
-        query = ' '.join((
-
-            "UPDATE path",
-            "SET strPath = ?, strContent = ?, strScraper = ?, noUpdate = ?",
-            "WHERE idPath = ?"
-        ))
-        kodicursor.execute(query, (path, None, None, 1, pathid))
-
-        # Update the file
-        query = ' '.join((
-
-            "UPDATE files",
-            "SET idPath = ?, strFilename = ?, dateAdded = ?",
-            "WHERE idFile = ?"
-        ))
-        kodicursor.execute(query, (pathid, filename, dateadded, fileid))
-        
-        # Process cast
-        people = artwork.get_people_artwork(item['People'])
-        self.kodi_db.addPeople(episodeid, people, "episode")
-        # Process artwork
-        artworks = artwork.get_all_artwork(item)
-        artwork.add_update_art(artworks['Primary'], episodeid, "episode", "thumb", kodicursor)
-        # Process stream details
-        streams = API.get_media_streams()
-        self.kodi_db.addStreams(fileid, streams, runtime)
-        # Process playstates
-        resume = API.adjust_resume(userdata['Resume'])
-        total = round(float(runtime), 6)
-        self.kodi_db.addPlaystate(fileid, resume, total, playcount, dateplayed)
-        if not self.directpath and resume:
-            # Create additional entry for widgets. This is only required for plugin/episode.
-            temppathid = self.kodi_db.getPath("plugin://plugin.video.emby.tvshows/")
-            tempfileid = self.kodi_db.addFile(filename, temppathid)
-            query = ' '.join((
-
-                "UPDATE files",
-                "SET idPath = ?, strFilename = ?, dateAdded = ?",
-                "WHERE idFile = ?"
-            ))
-            kodicursor.execute(query, (temppathid, filename, dateadded, tempfileid))
-            self.kodi_db.addPlaystate(tempfileid, resume, total, playcount, dateplayed)
-
-    def updateUserdata(self, item):
-        # This updates: Favorite, LastPlayedDate, Playcount, PlaybackPositionTicks
-        # Poster with progress bar
-        emby_db = self.emby_db
-        API = api.API(item)
-        
-        # Get emby information
-        itemid = item['Id']
-        checksum = API.get_checksum()
-        userdata = API.get_userdata()
-        runtime = API.get_runtime()
-        dateadded = API.get_date_created()
-
-        # Get Kodi information
-        emby_dbitem = emby_db.getItem_byId(itemid)
-        try:
-            kodiid = emby_dbitem[0]
-            fileid = emby_dbitem[1]
-            mediatype = emby_dbitem[4]
-            log.info(
-                "Update playstate for %s: %s fileid: %s"
-                % (mediatype, item['Name'], fileid))
-        except TypeError:
-            return
-
-        # Process favorite tags
-        if mediatype == "tvshow":
-            if userdata['Favorite']:
-                self.kodi_db.addTag(kodiid, "Favorite tvshows", "tvshow")
-            else:
-                self.kodi_db.removeTag(kodiid, "Favorite tvshows", "tvshow")
-        elif mediatype == "episode":
-            # Process playstates
-            playcount = userdata['PlayCount']
-            dateplayed = userdata['LastPlayedDate']
-            resume = API.adjust_resume(userdata['Resume'])
-            total = round(float(runtime), 6)
-
-            log.debug("%s New resume point: %s" % (itemid, resume))
-
-            self.kodi_db.addPlaystate(fileid, resume, total, playcount, dateplayed)
-            if not self.directpath and not resume:
-                # Make sure there's no other bookmarks created by widget.
-                filename = self.kodi_db.getFile(fileid)
-                self.kodi_db.removeFile("plugin://plugin.video.emby.tvshows/", filename)
-
-            if not self.directpath and resume:
-                # Create additional entry for widgets. This is only required for plugin/episode.
-                filename = self.kodi_db.getFile(fileid)
-                temppathid = self.kodi_db.getPath("plugin://plugin.video.emby.tvshows/")
-                tempfileid = self.kodi_db.addFile(filename, temppathid)
-                query = ' '.join((
-
-                    "UPDATE files",
-                    "SET idPath = ?, strFilename = ?, dateAdded = ?",
-                    "WHERE idFile = ?"
-                ))
-                self.kodicursor.execute(query, (temppathid, filename, dateadded, tempfileid))
-                self.kodi_db.addPlaystate(tempfileid, resume, total, playcount, dateplayed)
-
-        emby_db.updateReference(itemid, checksum)
-
-    def remove(self, itemid):
-        # Remove showid, fileid, pathid, emby reference
-        emby_db = self.emby_db
-        embycursor = self.embycursor
-        kodicursor = self.kodicursor
-        artwork = self.artwork
-
-        emby_dbitem = emby_db.getItem_byId(itemid)
-        try:
-            kodiid = emby_dbitem[0]
-            fileid = emby_dbitem[1]
-            pathid = emby_dbitem[2]
-            parentid = emby_dbitem[3]
-            mediatype = emby_dbitem[4]
-            log.info("Removing %s kodiid: %s fileid: %s" % (mediatype, kodiid, fileid))
-        except TypeError:
-            return
-
-        ##### PROCESS ITEM #####
-
-        # Remove the emby reference
-        emby_db.removeItem(itemid)
-
-
-        ##### IF EPISODE #####
-
-        if mediatype == "episode":
-            # Delete kodi episode and file, verify season and tvshow
-            self.removeEpisode(kodiid, fileid)
-
-            # Season verification
-            season = emby_db.getItem_byKodiId(parentid, "season")
-            try:
-                showid = season[1]
-            except TypeError:
-                return
-            
-            season_episodes = emby_db.getItem_byParentId(parentid, "episode")
-            if not season_episodes:
-                self.removeSeason(parentid)
-                emby_db.removeItem(season[0])
-
-            # Show verification
-            show = emby_db.getItem_byKodiId(showid, "tvshow")
-            query = ' '.join((
-
-                "SELECT totalCount",
-                "FROM tvshowcounts",
-                "WHERE idShow = ?"
-            ))
-            kodicursor.execute(query, (showid,))
-            result = kodicursor.fetchone()
-            if result and result[0] is None:
-                # There's no episodes left, delete show and any possible remaining seasons
-                seasons = emby_db.getItem_byParentId(showid, "season")
-                for season in seasons:
-                    self.removeSeason(season[1])
-                else:
-                    # Delete emby season entries
-                    emby_db.removeItems_byParentId(showid, "season")
-                self.removeShow(showid)
-                emby_db.removeItem(show[0])
-
-        ##### IF TVSHOW #####
-
-        elif mediatype == "tvshow":
-            # Remove episodes, seasons, tvshow
-            seasons = emby_db.getItem_byParentId(kodiid, "season")
-            for season in seasons:
-                seasonid = season[1]
-                season_episodes = emby_db.getItem_byParentId(seasonid, "episode")
-                for episode in season_episodes:
-                    self.removeEpisode(episode[1], episode[2])
-                else:
-                    # Remove emby episodes
-                    emby_db.removeItems_byParentId(seasonid, "episode")
-            else:
-                # Remove emby seasons
-                emby_db.removeItems_byParentId(kodiid, "season")
-
-            # Remove tvshow
-            self.removeShow(kodiid)
-
-        ##### IF SEASON #####
-
-        elif mediatype == "season":
-            # Remove episodes, season, verify tvshow
-            season_episodes = emby_db.getItem_byParentId(kodiid, "episode")
-            for episode in season_episodes:
-                self.removeEpisode(episode[1], episode[2])
-            else:
-                # Remove emby episodes
-                emby_db.removeItems_byParentId(kodiid, "episode")
-            
-            # Remove season
-            self.removeSeason(kodiid)
-
-            # Show verification
-            seasons = emby_db.getItem_byParentId(parentid, "season")
-            if not seasons:
-                # There's no seasons, delete the show
-                self.removeShow(parentid)
-                emby_db.removeItem_byKodiId(parentid, "tvshow")
-
-        log.info("Deleted %s: %s from kodi database" % (mediatype, itemid))
-
-    def removeShow(self, kodiid):
-        
-        kodicursor = self.kodicursor
-        self.artwork.delete_artwork(kodiid, "tvshow", kodicursor)
-        kodicursor.execute("DELETE FROM tvshow WHERE idShow = ?", (kodiid,))
-        log.debug("Removed tvshow: %s." % kodiid)
-
-    def removeSeason(self, kodiid):
-        
-        kodicursor = self.kodicursor
-
-        self.artwork.delete_artwork(kodiid, "season", kodicursor)
-        kodicursor.execute("DELETE FROM seasons WHERE idSeason = ?", (kodiid,))
-        log.debug("Removed season: %s." % kodiid)
-
-    def removeEpisode(self, kodiid, fileid):
-
-        kodicursor = self.kodicursor
-
-        self.artwork.delete_artwork(kodiid, "episode", kodicursor)
-        kodicursor.execute("DELETE FROM episode WHERE idEpisode = ?", (kodiid,))
-        kodicursor.execute("DELETE FROM files WHERE idFile = ?", (fileid,))
-        log.debug("Removed episode: %s." % kodiid)
 
 class Music(Items):
