+++ resources/lib/objects/__init__.py
@@ -1,2 +1,4 @@
 # Dummy file to make this directory a package.
 from movies import Movies
+from musicvideos import MusicVideos
+from tvshows import TVShows
